+++ src/graphql/netflow.rs
@@ -251,6 +251,7 @@ impl NetflowQuery {
         paged_events_in_cluster!(
             ctx,
             filter,
+            filter.into(),
             filter.source,
             after,
             before,
@@ -279,6 +280,7 @@ impl NetflowQuery {
         paged_events_in_cluster!(
             ctx,
             filter,
+            filter.into(),
             filter.source,
             after,
             before,